Extract the schema graph class from reorg.py for it to be reused.
